Changes the path of log file. Urges admin to run electrum server as non root.
Till now start script was uging /var/log/electrum.log as the log file path.
That's problematic, when electrum server is launched by an unprivileged user
and without sudo. As a matter of fact there is no need to run electrum server
as root, and suggested way in electrum server's HOWTO is creating a new user.

Instead of messing with /var/log permissions or running start script with
sudo, this commit changes logging directory to user's home.

Also it urges the user to run script as non root user.