Close connections manually
This is a required change to run electrum-server on PyPy.

On PyPy the garbage collector works differently than it does on CPython. This in turns causes the connections to stay open and to eat up all your sockets in no time.

This change closes the connections right after reading the data out of it.