Make IPv4 optional for peers
This change removes the necessarity for peers (e.g. other servers) to have an IPv4 address that the local DNS can resolve.

That's useful for Tor hidden services whose .onion addresses can't be resolved through the usual DNS system. We'll just leave it to the clients to resolve IP addresses.