In electrum-server, ensure typical bin directories are on $PATH.
When running an init script at boot, the $PATH variable may not
include some typical directories like /usr/local/bin. This change
ensures those directories are on the path, so that bash can find
run_electrum_server.py if it is /usr/local/bin or one of the
other typical bin directories.